CM3: Conditionally run extra DDR steps based on command version
As part of the mainline work, some of the suspend steps were moved
down into the firmware. To support both kernels, add a V2 command
where the extra DDR steps are run. This allows the firmware to be
backwards compatible, but needs a new kernel that runs the new
command versions.

Signed-off-by: Russ Dill <Russ.Dill@ti.com>